pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Machine Learning RAG Assistant"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-google-genai",
    "langchain-community",
    "faiss-cpu",
    "pypdf",
    "python-dotenv",
    "streamlit",
]
